fixes in integration tests

+++ tests/java/integration/src/test/java/org/netxms/tests/ScriptTest.java
@@ -21,7 +21,6 @@ package org.netxms.tests;
 import java.net.InetAddress;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.UUID;
 import org.apache.commons.io.IOUtils;
 import org.netxms.base.InetAddressEx;
 import org.netxms.base.MacAddress;
@@ -42,6 +41,7 @@ public class ScriptTest extends AbstractSessionTest implements TextOutputListene
 {
    private NXCSession session;
    private Boolean scriptFailed;
+   private String currentScriptName = "<unset>";
    
    public void testAddressMap() throws Exception
    {
@@ -66,11 +66,13 @@ public class ScriptTest extends AbstractSessionTest implements TextOutputListene
       session.disconnect();
    }
    
-   private void executeScript(String script, List <String> params) throws Exception
+   private void executeScript(String resourceName, List<String> params) throws Exception
    {
+      currentScriptName = resourceName;
+      String script = IOUtils.toString(this.getClass().getResourceAsStream(resourceName), "UTF-8");
       ScriptCompilationResult r = session.compileScript(script, true);
       if (r.errorMessage != null)
-         System.out.println("Compilation error message: \"" + r.errorMessage + "\"");
+         System.out.println("Script " + resourceName + " compilation error: \"" + r.errorMessage + "\"");
       assertTrue(r.success);
       assertNotNull(r.code);
       assertNull(r.errorMessage);
@@ -83,7 +85,7 @@ public class ScriptTest extends AbstractSessionTest implements TextOutputListene
    public void testNXSLObjectFunctions() throws Exception
    {
       session = connect();
-      String script = IOUtils.toString(this.getClass().getResourceAsStream("/objectFunctions.nxsl"), "UTF-8");      
+      session.setCommandTimeout(120000);
 
       session.syncObjects();
       AbstractObject object = session.findObjectById(2);
@@ -95,54 +97,61 @@ public class ScriptTest extends AbstractSessionTest implements TextOutputListene
       long nodeId = session.createObject(cd);
       assertFalse(nodeId == 0);
       
-      object = session.findObjectById(3);
-      assertNotNull(object);      
+      long templateId = 0;
+      try
+      {
+         object = session.findObjectById(3);
+         assertNotNull(object);
 
-      cd = new NXCObjectCreationData(AbstractObject.OBJECT_TEMPLATE, "TestTemplate", 3);
-      long templateId = session.createObject(cd);
-      assertFalse(templateId == 0);
+         cd = new NXCObjectCreationData(AbstractObject.OBJECT_TEMPLATE, "TestTemplate", 3);
+         templateId = session.createObject(cd);
+         assertFalse(templateId == 0);
 
-      Thread.sleep(1000);  // Object update should be received from server
+         Thread.sleep(1000); // Object update should be received from server
 
-      object = session.findObjectById(nodeId);
-      assertNotNull(object);
-      object = session.findObjectById(templateId);
-      assertNotNull(object);  
-      
-      session.applyTemplate(templateId, nodeId);
+         object = session.findObjectById(nodeId);
+         assertNotNull(object);
+         object = session.findObjectById(templateId);
+         assertNotNull(object);
 
-      String interfaceName1 = "lo";
-      cd = new NXCObjectCreationData(AbstractObject.OBJECT_INTERFACE, interfaceName1, nodeId);
-      cd.setMacAddress(MacAddress.parseMacAddress("00:10:FA:23:11:7A"));
-      cd.setIpAddress(new InetAddressEx(InetAddress.getLoopbackAddress(), 0));
-      cd.setIfType(24);
-      cd.setIfIndex(0);
-      session.createObject(cd);
+         session.applyTemplate(templateId, nodeId);
 
-      String interfaceName2 = "eth0";
-      cd = new NXCObjectCreationData(AbstractObject.OBJECT_INTERFACE, interfaceName2, nodeId);
-      cd.setMacAddress(MacAddress.parseMacAddress("01 02 fa c4 10 dc"));
-      cd.setIpAddress(new InetAddressEx(InetAddress.getByName("192.168.10.1"), 24));
-      cd.setIfType(6);
-      cd.setIfIndex(1);
-      session.createObject(cd);
-      
-      List<String> params = new ArrayList<String>();
-      params.add(Long.toString(nodeId));
-      params.add(Long.toString(templateId));
-      params.add(interfaceName1);
-      params.add(interfaceName2);
-      executeScript(script, params);
-      
-      session.deleteObject(nodeId);
-      session.deleteObject(templateId);      
-      session.disconnect();
+         String interfaceName1 = "lo";
+         cd = new NXCObjectCreationData(AbstractObject.OBJECT_INTERFACE, interfaceName1, nodeId);
+         cd.setMacAddress(MacAddress.parseMacAddress("00:10:FA:23:11:7A"));
+         cd.setIpAddress(new InetAddressEx(InetAddress.getLoopbackAddress(), 0));
+         cd.setIfType(24);
+         cd.setIfIndex(0);
+         session.createObject(cd);
+
+         String interfaceName2 = "eth0";
+         cd = new NXCObjectCreationData(AbstractObject.OBJECT_INTERFACE, interfaceName2, nodeId);
+         cd.setMacAddress(MacAddress.parseMacAddress("01 02 fa c4 10 dc"));
+         cd.setIpAddress(new InetAddressEx(InetAddress.getByName("192.168.10.1"), 24));
+         cd.setIfType(6);
+         cd.setIfIndex(1);
+         session.createObject(cd);
+
+         List<String> params = new ArrayList<String>();
+         params.add(Long.toString(nodeId));
+         params.add(Long.toString(templateId));
+         params.add(interfaceName1);
+         params.add(interfaceName2);
+         executeScript("/objectFunctions.nxsl", params);
+      }
+      finally
+      {
+         if (nodeId != 0)
+            session.deleteObject(nodeId);
+         if (templateId != 0)
+            session.deleteObject(templateId);
+         session.disconnect();
+      }
    }
-   
+
    public void testNXSLDataCollectionFunctions() throws Exception
    {
       session = connect();
-      String script = IOUtils.toString(this.getClass().getResourceAsStream("/dataCollectionFunctions.nxsl"), "UTF-8");      
 
       session.syncObjects();
       List<AbstractObject> objects = session.getAllObjects(); //Find managment node
@@ -161,7 +170,7 @@ public class ScriptTest extends AbstractSessionTest implements TextOutputListene
       List<String> params = new ArrayList<String>();
       params.add(Long.toString(managementNode.getObjectId()));
       params.add(dciName);
-      executeScript(script, params);
+      executeScript("/dataCollectionFunctions.nxsl", params);
       
       DataCollectionConfiguration config = session.openDataCollectionConfiguration(managementNode.getObjectId());
       for (DataCollectionObject dc : config.getItems())
@@ -179,7 +188,6 @@ public class ScriptTest extends AbstractSessionTest implements TextOutputListene
    public void testNXSLAgentFunctions() throws Exception
    {
       session = connect();
-      String script = IOUtils.toString(this.getClass().getResourceAsStream("/agentFunctions.nxsl"), "UTF-8");      
 
       session.syncObjects();
       List<AbstractObject> objects = session.getAllObjects(); //Find managment node
@@ -197,42 +205,46 @@ public class ScriptTest extends AbstractSessionTest implements TextOutputListene
       String actionName = TestConstants.ACTION;
       String actionName2 = "testEcho";
 
-      NXCObjectCreationData cd = new NXCObjectCreationData(AbstractObject.OBJECT_TEMPLATE, "TestTemplate", 3);
+      NXCObjectCreationData cd = new NXCObjectCreationData(AbstractObject.OBJECT_TEMPLATE, "ScriptTestTemplate", 3);
       long templateId = session.createObject(cd);
       assertFalse(templateId == 0);
 
-      Thread.sleep(1000);  // Object update should be received from server
+      try
+      {
+         Thread.sleep(1000); // Object update should be received from server
 
-      AbstractObject object = session.findObjectById(templateId);
-      assertNotNull(object); 
+         AbstractObject object = session.findObjectById(templateId);
+         assertNotNull(object);
 
-      AgentPolicy policy = new AgentPolicy("TestPolicy", AgentPolicy.AGENT_CONFIG);
-      policy.setContent("Action = " + actionName2 + ": echo $1 $2 $3 $4");
-      session.savePolicy(templateId, policy, false);
+         AgentPolicy policy = new AgentPolicy("TestPolicy", AgentPolicy.AGENT_CONFIG);
+         policy.setContent("Action = " + actionName2 + ": echo $1 $2 $3 $4");
+         session.savePolicy(templateId, policy, false);
 
-      session.applyTemplate(templateId, managementNode.getObjectId());    
-      session.executeActionWithExpansion(managementNode.getObjectId(), 0, "Agent.Restart", false, null, null, null);
+         session.applyTemplate(templateId, managementNode.getObjectId());
+         session.executeActionWithExpansion(managementNode.getObjectId(), 0, "Agent.Restart", false, null, null, null);
 
-      Thread.sleep(10000);  // Wait for agent restart
-      
-      List<String> params = new ArrayList<String>();
-      params.add(Long.toString(managementNode.getObjectId()));
-      params.add(actionName);
-      params.add(actionName2);
-      executeScript(script, params);
-      
-      session.deleteObject(templateId);
-      
-      session.disconnect();
+         Thread.sleep(10000); // Wait for agent restart
+
+         List<String> params = new ArrayList<String>();
+         params.add(Long.toString(managementNode.getObjectId()));
+         params.add(actionName);
+         params.add(actionName2);
+         executeScript("/agentFunctions.nxsl", params);
+      }
+      finally
+      {
+         session.deleteObject(templateId);
+         session.disconnect();
+      }
    }
 
-   /* (non-Javadoc)
+   /**
     * @see org.netxms.client.ActionExecutionListener#messageReceived(java.lang.String)
     */
    @Override
    public void messageReceived(final String text)
    {
-      System.out.println(text);
+      System.out.println("[" + currentScriptName + "] " + text);
    }
 
    @Override
@@ -243,7 +255,7 @@ public class ScriptTest extends AbstractSessionTest implements TextOutputListene
    @Override
    public void onError()
    {
-      System.out.println("Script error occured");
+      System.out.println("[" + currentScriptName + "] EXECUTION ERROR");
       scriptFailed = true;
    }
 }
